Always match correct tsi on erasure (RhBug:1253120)
Since the RPM callbacks require a TransactionItem instance but they only
receive the package name on erasures (as opposed to installs), we need
to look for the respective tsi ourselves.  This was done by using the
provided name but that didn't work correctly when multiple packages of
the same name were being erased (e.g. multiple kernels); the lookup
would always pick the first matching tsi that appeared in the
transaction, ignoring the others.

This commit fixes that by making use of the _current_te attribute (added
by commit dfede4c) which provides the package NEVRA instead, allowing us
to match the correct tsi.