Always match correct tsi on erasure (RhBug:1253120)
RPM callbacks only receive the package name on erasures.  However, we
need the package object, so we have to look it up ourselves.  This was
done by using the provided name but that didn't work correctly when
multiple packages of the same name were being erased (e.g. multiple
kernels); the lookup would always pick the first matching tsi that
appeared in the transaction, ignoring the others.

This commit fixes that by making use of the new
RPMCALLBACK_ELEM_PROGRESS type introduced to RPM by the commit baec562,
allowing us to use NEVRA in the lookup instead.