Make the binaries at build time
Instead of having separate copies of the dnf and dnf-automatic binaries
for each python version, let's just generate those at build time using
.in files.

Note that the set of files generated at build time remains the same as
before this commit, so no change in the .spec file is required.

Since the binaries are now placed in CMAKE_BINARY_DIR, we must ensure
that, for out of source builds, they still continue to work for
developers without installing.  To do that, we need to configure the
dnf/__init__.py file so that it's placed in CMAKE_BINARY_DIR and extend
its __path__ attribute to include CMAKE_SOURCE_DIR.  To make sure the
configuration doesn't happen with builds meant to be installed, the
CHECKOUT cmake option has been introduced.